requirements-root.txt
django
psycopg2-binary
python-dotenv 
pytest
tox
isort
flake8
pytest-django
pytest-flake8
pytest-isort
pytest-cov
datapunt-django-snapshot

